Update player components and spawning
These components should make it easier to do queries for specifically
these components when we need to do computations and updates on them.

Because it doesn't seem like you can query for outer bundles, I'm adding
a player component as part of a new PlayerBundle, mostly for ease of
querying.

Maybe this will change in a future version of bevy, but this is fine
for now.