Add systems for inputs and movement
This sets a speed based on what key is being set, and then also sets
the speed to 0 if nothing is pressed for those buttons.

That code can be cleaned up for sure, but it works for now. This speed
is maybe slow, but I still need to figure out world coordinates for the
game.