Add new systems and ensure ordering
The labels I'm adding to input and player movement help to make sure
that the systems run in the right order. Normally bevy will attempt to
run systems in parallel as much as possible, but I want to ensure that
the speed has been updated _before_ I compute where the player is at.