Swap to integer speed
It looks like the tile map crate I'll be using works best with integer
speeds as you move the character around the map (we'll see what happens
when I get into the physics portion). Because of the values I was using
this doesn't really make any appreciable difference but I want to do it
now instead of later.