Use a grid for drawing background
Previously when drawing the tile I was getting overlaps when drawing
the tile because I was just plopping the tile at the player's x,y
coordinate. By instead tying it to a grid cell it means no more
overlaps.

I hope to eventually not have to iterate through all of the tiles to
find the intersecting ones, but for now it is fine.

The only major problem with this is that I'm generating the entire grid
up front, and if the player moves out of the grid nothing happens. I'd
like to eventually change this so that the game generates new cells as
the player moves, and play around a bit more with that idea.

I'm leaving in some commented code for now that draws an outline. When
the grid is large this makes it chug so I'm not keeping it forever, but
it is good for debugging sometimes.