Rename struct and use indexes
When I was storing the handle I would have to constantly get the actual
texture atlas out of all of the texture atlases. However if I store the
index instead, I only need that and the handle for the texture atlas
when spawning new sprites, so it makes everything a bit cleaner.