Grid movement and camera zoom
Now we have the player movement snapping to a grid by ensuring that the
speed is the same as the grid size. I've also put in some code to allow
for small presses of the key to move 1 grid cell, and also hopefully
allow for continuous movement.

I'm also zooming in the camera a bit so that you see less of the map,
with the possibility of zooming the camera in and out later.