Spawn crop components
Now we are spawning actual crop bundles inside of our actions, and
instead of always just blindly spawning one, I only spawn a crop if
there isn't already a crop at the same location. Eventually I might
link the crop to the cell directly as well to maybe make this a bit
easier to work with.