Add module for game world
Right now this only contains the timer for a world "tick". A world
"tick" is I think similar to how Minecraft also only calculates things
every so often. I don't want to calculate crops growing or dying every
time my system updates, as this could come in as fast as the computer
can send it. Instead I want to have a fairly steady stream of the world
essentially calculating all of the various entities inside of it. This
would be separate from things like animation or physics and has more to
with things in the background.