Add more to crop component and update spawning
Because we now have stages I need to copy over more of that
configuration into my component so that it's easier to advance the
crop growth and change the sprite to be the next one in the line. I
might look into getting the crop configuration from the app, but I think
this works for right now.