Add crop growth system
Every time the world tick rolls over, I "grow" the crop by incrementing
the number of ticks the crop has been in the stage. Once the crop hits
the minimum number of time it needs to be in the stage, it _might_
advance to the next stage if the random number is in the right range.