Revamp our grid cell
I don't think the way I was doing this was very smart, so I wanted to
revamp it. Now I can just flip a visibility flag to get the sprite to
be drawn, and I'll be able to use these entities from within other
systems without having to go through the Grid struct.

With this update the Grid became useless so I can just get rid of it.