Fix bug with camera scale when leaving inventory
Because I'm removing and adding back in the Orthographic camera I
forgot that the user could have zoomed the camera to a different scale
to what it was when the game launched. So what I'm doing is storing
the scale of the camera separate from the actual camera itself, just so
that when I come back to spawn it again it's at the right scale.