Make look-ahead for visibility
Now that I know the direction that the player is moving, I can make the
squares that are in front of them visible as well as the square they
are currently on.

I'm also adding in a PlayerMovement struct to keep all of that type of
information about the player. It will be used in the future.