Start world generation
This is the start of my world generation, all smooshed together into a
single commit. It has the start of a bad "room template" which is just
a horrible function that wouldn't scale whatsoever. However I think
this might be something that I'll at least do something similar going
forward.

There's also now a grid which has simpler coordinates than our normal
game coordinates that include the tile size. This can set the cell type
for each of the cells inside of it, and then the spawning can read those
to know what sprite to use and whether it is a wall or not.