Spawn map tiles from from grid
This does clean this method up quite a bit now because I don't need to
run through a bunch of loops to generate my tiles. Now I can just read
the information from the grid and spawn what I need to. I still need to
find a way to not nest these as much as they are.