Add more room templates and randomly pick
Now I have multiple templates and I can randomly pick one. In the future
I might want to give each size of template a chance to be chosen instead
of an equal chance of any of them, but this is fine, it just means the
bigger rooms have a bigger chance of being chosen.