Add flag for setting textures
This will make it easier for our "loading" systems to know when to move
into the next app state.

I'm going to move more of the state switching logic into one place so
it is easier to see all of the state transitions and their requirements
so they aren't scattered across disparate systems in disparate files.