Generate world during loading
I don't need to generate this during the initial spawn of the world
entities, instead I can generate the world up front, and then just
spawn things from the grid that is inside of the resources.

This will eventually let me tie into the asset server to maybe load
room templates from files, or to load a saved grid instead.