Players can break walls
Now when a player moves into a wall, if the wall can be broken and
they are holding a tool that can break walls the wall gets broken and
they can move into it. As part of this I need to despawn the wall and
spawn in the broken wall.