Add spawns component
This will give me a single point to push information any time I want to
spawn some new entities. This way when I want to spawn a crop, I do
not need to have that same code all over the codebase, but it is instead
only in the 1 system.

Even though I was previously only spawning crops in 1 place I was about
to add some functionality so that when crops die, they drop their seeds
all over again.