Remove all sprite numbers
Other than the defaults which only live where the files are loaded
everything else is now grabbed from config, which means it all comes
through that load instead of it just being a const in the code.

I probably won't change any of those things, but it's still nice to
not really have those values in the code.