Add item spawning
Now that I have a few tools, I can spawn one of them in the world so
that it can later be picked up. Right now it's just hard-coded as a
shovel, but in the future I'll add more items later.

This also updates a few things in the other bundle spawning.