Tools can clear structures
This is different than destroying, as it just completely removes them
from the game. The idea is that for things like dirt clumps or rubble,
it could be "tilled" back into the soil or something, maybe eventually
changing up the look of the cell, but right now it just puts it back
to what it was.