Move map spawning to system
Instead of spawning from a map in 2 places, I've moved it all into a
single system. Unfortunately as part of this it means I can't assume
that the player will be there in all of the other systems. I might
need to rework this a bit eventually so that the spawns system runs
in both the enter in game and on update.