Update world ticking and some spawns
This moves our world ticking into the world plugin, which probably
makes the most sense, and creates a new component for that inside of
the plugin that the main app can use in queries. The new component just
specifies if a tick has just completed.

I'm also moving the spawning of the gameplay camera to the add system,
so that the initial_spawns system can be trimmed down even more.